fix: Moved user settings to separate page
